Fix login

+++ Entidades/Cliente.cs
@@ -8,7 +8,7 @@ namespace Entidades
 {
     public class Cliente
     {
-        public int Dni { get; set; }
+        public long Dni { get; set; }
         public string Apellido { get; set; }
         public string Nombre { get; set; }
         public Usuario Usuario { get; set; }
